# [various] Speed up build-examples