--- a/packages/animations/example/android/.pluginToolsConfig.yaml
+++ b/packages/animations/example/android/.pluginToolsConfig.yaml
@@ -1,0 +1,3 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"