--- a/packages/animations/example/android/.pluginToolsConfig.yaml
+++ b/packages/animations/example/android/.pluginToolsConfig.yaml
@@ -1,3 +1,4 @@
 buildFlags:
   _pluginToolsConfigGlobalKey:
     - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"