--- a/packages/camera/camera_android_camerax/example/.pluginToolsConfig.yaml
+++ b/packages/camera/camera_android_camerax/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"