--- a/packages/dynamic_layouts/example/.pluginToolsConfig.yaml
+++ b/packages/dynamic_layouts/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"