--- a/packages/extension_google_sign_in_as_googleapis_auth/example/.pluginToolsConfig.yaml
+++ b/packages/extension_google_sign_in_as_googleapis_auth/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"