--- a/packages/flutter_adaptive_scaffold/example/.pluginToolsConfig.yaml
+++ b/packages/flutter_adaptive_scaffold/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"