--- a/packages/flutter_plugin_android_lifecycle/example/.pluginToolsConfig.yaml
+++ b/packages/flutter_plugin_android_lifecycle/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"