--- a/packages/google_maps_flutter/google_maps_flutter/example/.pluginToolsConfig.yaml
+++ b/packages/google_maps_flutter/google_maps_flutter/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"