--- a/packages/google_sign_in/google_sign_in/example/.pluginToolsConfig.yaml
+++ b/packages/google_sign_in/google_sign_in/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"