--- a/packages/local_auth/local_auth_android/example/.pluginToolsConfig.yaml
+++ b/packages/local_auth/local_auth_android/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"