--- a/packages/palette_generator/example/.pluginToolsConfig.yaml
+++ b/packages/palette_generator/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"