--- a/packages/path_provider/path_provider_android/example/.pluginToolsConfig.yaml
+++ b/packages/path_provider/path_provider_android/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"