--- a/packages/pigeon/example/app/.pluginToolsConfig.yaml
+++ b/packages/pigeon/example/app/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"