--- a/packages/quick_actions/quick_actions/example/.pluginToolsConfig.yaml
+++ b/packages/quick_actions/quick_actions/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"