--- a/packages/rfw/example/remote/.pluginToolsConfig.yaml
+++ b/packages/rfw/example/remote/.pluginToolsConfig.yaml
@@ -1,3 +1,4 @@
 buildFlags:
-  global:
+  _pluginToolsConfigGlobalKey:
     - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"