--- a/packages/rfw/example/remote/.pluginToolsConfig.yaml
+++ b/packages/rfw/example/remote/.pluginToolsConfig.yaml
@@ -1,4 +1,0 @@
-buildFlags:
-  _pluginToolsConfigGlobalKey:
-    - "--no-tree-shake-icons"
-    - "--dart-define=buildmode=testing"