--- a/packages/rfw/example/remote/.pluginToolsConfig.yaml
+++ b/packages/rfw/example/remote/.pluginToolsConfig.yaml
@@ -1,0 +1,3 @@
+buildFlags:
+  global:
+    - "--no-tree-shake-icons"