--- a/packages/rfw/example/wasm/.pluginToolsConfig.yaml
+++ b/packages/rfw/example/wasm/.pluginToolsConfig.yaml
@@ -1,0 +1,3 @@
+buildFlags:
+  global:
+    - "--no-tree-shake-icons"