--- a/packages/shared_preferences/shared_preferences/example/.pluginToolsConfig.yaml
+++ b/packages/shared_preferences/shared_preferences/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"