--- a/packages/url_launcher/url_launcher/example/.pluginToolsConfig.yaml
+++ b/packages/url_launcher/url_launcher/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"