--- a/packages/video_player/video_player/example/.pluginToolsConfig.yaml
+++ b/packages/video_player/video_player/example/.pluginToolsConfig.yaml
@@ -1,0 +1,4 @@
+buildFlags:
+  _pluginToolsConfigGlobalKey:
+    - "--no-tree-shake-icons"
+    - "--dart-define=buildmode=testing"