--- a/script/tool/lib/src/common/package_state_utils.dart
+++ b/script/tool/lib/src/common/package_state_utils.dart
@@ -181,8 +181,6 @@ Future<bool> _isDevChange(List<String> pathComponents,
       pathComponents.contains('lint-baseline.xml') ||
       // Example build files are very unlikely to be interesting to clients.
       _isExampleBuildFile(pathComponents) ||
-      // Build flags are very unlikely to be interesting to clients.
-      pathComponents.contains('.pluginToolsConfig.yaml') ||
       // Test-only gradle depenedencies don't affect clients.
       await _isGradleTestDependencyChange(pathComponents,
           git: git, repoPath: repoPath);
@@ -197,6 +195,7 @@ bool _isExampleBuildFile(List<String> pathComponents) {
       pathComponents.contains('build.gradle') ||
       pathComponents.contains('Runner.xcodeproj') ||
       pathComponents.contains('CMakeLists.txt') ||
+      pathComponents.contains('.pluginToolsConfig.yaml') ||
       pathComponents.contains('pubspec.yaml');
 }