https://github.com/flutter/flutter/issues/125707

- Create pluginTools config for every example that shrunk material icons and had an android config

Likely some build caching but before/after If someone knows how to clean the cache or force a rebuild I will give more accurate after numbers. 
```
dart run ./script/tool/bin/flutter_plugin_tools.dart build-examples --apk  66.13s user 15.36s system 8% cpu 15:16.13 total
after 
dart run ./script/tool/bin/flutter_plugin_tools.dart build-examples --apk  52.24s user 11.65s system 50% cpu 2:07.10 total
```
## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [x] I read the [Tree Hygiene] wiki page, which explains my responsibilities.
- [x] I read and followed the [relevant style guides] and ran the auto-formatter. (Unlike the flutter/flutter repo, the flutter/packages repo does use `dart format`.)
- [x] I signed the [CLA].
- [x] The title of the PR starts with the name of the package surrounded by square brackets, e.g. `[shared_preferences]`
- [x] I listed at least one issue that this PR fixes in the description above.
- [x] I updated `pubspec.yaml` with an appropriate new version according to the [pub versioning philosophy], or this PR is [exempt from version changes].
- [ ] I updated `CHANGELOG.md` to add a description of the change, [following repository CHANGELOG style].
- [ ] I updated/added relevant documentation (doc comments with `///`).
- [x] I added new tests to check the change I am making, or this PR is [test-exempt].
- [x] All existing and new tests are passing.