@Hixie 2023-05-01
test exemption should not be necessary, the PR modifies a test